tacky: refactoring to put switch/if in single file.
